+++ CHANGELOG.md
@@ -49,3 +49,4 @@ Created on Dec 20, 2023
 - Beta Release: Delete duplicated button on about page on May 19, 2024 -> __0.7761__
 - Beta Release: Add 'Dividends' transaction type option on July 2, 2024 -> __0.777__
 - Beta Release: Add 'Interests' transaction type option on July 2, 2024 -> __0.7771__
+- Beta Release: Add internal processes for newly added options on July 2, 2024 -> __0.7773__